Refactor, rearrange files
