--- packages/pretty-reporter/src/types.ts
@@ -1,7 +1,0 @@
-import type { Diagnostic, Severity } from '@steiger/types'
-
-export interface FullDiagnostic extends Diagnostic {
-  ruleName: string
-  severity: Exclude<Severity, 'off'>
-  ruleDescriptionUrl: string
-}
