+++ packages/steiger-plugin-fsd/src/excessive-slicing/index.ts
@@ -1,6 +1,6 @@
 import { join } from 'node:path'
 import { getLayers, getSlices, isSliced } from '@feature-sliced/filesystem'
-import type { Diagnostic, Rule } from '@steiger/types'
+import type { PartialDiagnostic, Rule } from '@steiger/types'
 
 import { groupSlices } from '../_lib/group-slices.js'
 import { NAMESPACE } from '../constants.js'
@@ -16,7 +16,7 @@ const THRESHOLDS = {
 const excessiveSlicing = {
   name: `${NAMESPACE}/excessive-slicing`,
   check(root) {
-    const diagnostics: Array<Diagnostic> = []
+    const diagnostics: Array<PartialDiagnostic> = []
 
     for (const [layerName, layer] of Object.entries(getLayers(root))) {
       if (!isSliced(layer) || !(layerName in THRESHOLDS)) {
