+++ packages/steiger-plugin-fsd/src/no-processes/index.ts
@@ -1,11 +1,11 @@
 import { basename } from 'node:path'
-import type { Diagnostic, Rule } from '@steiger/types'
+import type { PartialDiagnostic, Rule } from '@steiger/types'
 import { NAMESPACE } from '../constants.js'
 
 const noProcesses = {
   name: `${NAMESPACE}/no-processes`,
   check(root) {
-    const diagnostics: Array<Diagnostic> = []
+    const diagnostics: Array<PartialDiagnostic> = []
 
     const processesLayer = root.children.find(
       (child) => child.type === 'folder' && basename(child.path) === 'processes',
