+++ packages/steiger-plugin-fsd/src/shared-lib-grouping/index.ts
@@ -1,5 +1,5 @@
 import { getLayers, getSegments } from '@feature-sliced/filesystem'
-import type { Diagnostic, Rule } from '@steiger/types'
+import type { PartialDiagnostic, Rule } from '@steiger/types'
 import { NAMESPACE } from '../constants.js'
 
 const THRESHOLD = 15
@@ -8,7 +8,7 @@ const THRESHOLD = 15
 const sharedLibGrouping = {
   name: `${NAMESPACE}/shared-lib-grouping`,
   check(root) {
-    const diagnostics: Array<Diagnostic> = []
+    const diagnostics: Array<PartialDiagnostic> = []
 
     const { shared } = getLayers(root)
     if (!shared) {
