+++ packages/steiger/src/features/run-rule/types.ts
@@ -1,0 +1,9 @@
+import { File, Folder, Severity } from '@steiger/types'
+
+export interface SeverityMarkedFile extends File {
+  severity: Severity
+}
+
+export interface SeverityMarkedFolder extends Folder {
+  severity: Severity
+}
